+++ src-tauri/src/internal/dbo/mod.rs
@@ -1,1 +1,2 @@
+pub mod entry;
 pub mod feed; // Declare the rss module
